fix(stella-cli): split the steering switch into settings/steering.rs to clear the file-size ratchet
settings.rs crossed the 1500-line ceiling (1542) with the #3243 Phase 2
additions and is not grandfathered. Move SteeringCeiling,
Settings::steering_enabled, and env_toggle into a sibling submodule
(the settlement.rs pattern), and take clippy's is_none_or on the moved
derivation. settings.rs is back to 1485 lines; no behavior change.

Refs #3243
